# api/docs/swagger.yaml
basePath: /v1
definitions:
  http.Response:
    properties:
      data: {}
      description:
        type: string
      status:
        type: string
    type: object
...
  models.ResponseMessage:
    properties:
      message:
        type: string
    type: object
  models.ServicesRequest:
    properties:
      clinic_id:
        type: integer
      name:
        type: string
      price:
        type: number
    type: object
  models.ServicesResponse:
    properties:
      clinic_id:
        type: integer
      id:
        type: integer
      name:
        type: string
      price:
        type: number
    type: object
...
  models.UpdateServicesReq:
    properties:
      id:
        type: integer
      name:
        type: string
      price:
        type: number
    type: object
info:
  contact: {}
  description: This is Clinic server api
  termsOfService: 1 term Traffic Light
  title: Clinic
  version: "1.0"
paths:
...
  /service:
    post:
      consumes:
      - application/json
      description: Create Service
      operationId: createservice
      parameters:
      - description: CreateServiceInfo
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.ServicesRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Service data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.ServicesResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Create Service
      tags:
      - Service
    put:
      consumes:
      - application/json
      description: PUT Services
      operationId: putservices
      parameters:
      - description: UpdateServices
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.UpdateServicesReq'
      produces:
      - application/json
      responses:
        "200":
          description: Service data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.ServicesResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: PUT Services
      tags:
      - Service
  /service/{id}:
    get:
      consumes:
      - application/json
      description: Get Service
      operationId: getservicebyid
      parameters:
      - description: service_id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Service data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.ServicesResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get Service
      tags:
      - Service
  /service/delete/{id}:
    delete:
      consumes:
      - application/json
      description: PUT Service
      operationId: service
      parameters:
      - description: serviceID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Service data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.ResponseMessage'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: PUT Service
      tags:
      - Service
  /services:
    get:
      consumes:
      - application/json
      description: Get all Service
      operationId: getservices
      produces:
      - application/json
      responses:
        "200":
          description: Service data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.ServicesResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get all Service
      tags:
      - Service
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"